# lib/api-zod/src/generated/types/updateEmployeeBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SOMELEC Mission Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateEmployeeBody {
  /** @nullable */
  firstName?: string | null;
  /** @nullable */
  lastName?: string | null;
  /** @nullable */
  matricule?: string | null;
  /** @nullable */
  nni?: string | null;
  /** @nullable */
  position?: string | null;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  departmentId?: number | null;
}